Interim commit, preliminary code to make AR6 vetting page
The current code is non-functional and will throw errors. It creates tabs for
displaying pass/fail status and values, and filling them with a DataFrame table
each, but the IAM COMPACT-specific code required to produce those DataFrames
still needs to be finished in the `iamcompact-vetting` package. Buttons or other
widgets also need to be added to download results as Excel. The descriptions tab
is meant to describe each vetting check, but currently only displays an "Under
construction" message.