Copying AR6 vetting file to population/GDP vetting file.
Starting to write a new version of the population/GDP harmonization vetting
page, using the AR6 vetting page as a basis. This version is being commited to
be able to track which changes needed to be made to turn the page into a
vetting page for the population/GDP harmonization, which might then serve as a
basis for making a template or a class that can more easily be configured for
different vetting checks.