Defined name validation pages through explicit functions.
Apparently, even though the documentation of `streamlit.Page` states that it
accepts a "callable" to call for running a page rather than a path to a .py
code file, it appears to do some boneheaded processing that probably only works
with regular def functions, and not with partials or lambdas. It appears to not
work if the object you pass doesn't have a __name__ attribute, although the
real issue might be something else, didn't have time to investigate thoroughly.