Now recognizes model-native region names.
Validates region names using iamcompact_nomenclature.get_invalid_model_regions,
which checks the RegionProcessor for model-native region names, and tests per
model and region whether it is a valid combination. The details page for the
region name validation now displays a table with a column that states for each
invalid region which models use it without having it as a valid native region
(which will be all models in the data for regions that are not recognized
model-native regions for any model).