Fixed wrong label in vetting check status msg.
A mistake in a function to display status messages for vetting checks caused
all of them to be called "AR6 checks".

The same function should maybe be customized to display a customizable label.
Currently we use a single common function to display a status message for each
vetting check page, and just say that "checks" in general passed or failed.